introduce places.js helper file
